requirements.txt
# Tally ERP Multi-Agent System Dependencies
# Pure ReAct Agents 

# HTTP Requests for Tally
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0

# XML Parsing
lxml>=5.0.0

# Google Gemini AI
google-generativeai>=0.3.0

# LangChain Framework
langchain>=0.1.0
langchain-core>=0.1.0

# Google Gemini Integration for LangChain
langchain-google-genai>=1.0.0

# LangChain Hub for Prompt Templates
langchainhub>=0.1.14

# Additional dependencies
pydantic>=2.0.0
typing-extensions>=4.0.0

# Flask API
flask>=2.3.0
flask-cors>=4.0.0

# ODBC Driver
pyodbc>=5.0.0


matplotlib 
seaborn 
pillow
streamlit>=1.26.0
gradio

matplotlib
pillow

# --- API & Web Server (For the "Other Guy") ---
fastapi
uvicorn
pydantic


# --- Vector Database & Embeddings ---
chromadb
sentence-transformers

